requirements.txt
ultralytics 
opencv-python 
pillow 
pdf2image 
music21
label-studio
matplotlib